+++ R/compute_weight_gp.R
@@ -47,7 +47,10 @@
 #' kernel_fn <- function(x) exp(-x^2)
 #'
 #' # Estimate GPS function
-#' GPS_m <- train_GPS(cov_mt = data[,-(1:2)], w_all = data$treat)
+#' GPS_m <- train_gps(cov_mt = data[,-(1:2)],
+#'                    w_all = data$treat,
+#'                    sl_lib = c("SL.xgboost"),
+#'                    dnorm_log = FALSE)
 #'
 #' GPS <- GPS_m$GPS
 #'