+++ R/train_GPS.R
@@ -1,13 +1,14 @@
 #' @title
-#' Train A Model for GPS
+#' Train A Model for Generalized Propensity Score
 #'
 #' @description
 #' Estimates the conditional mean and sd of exposure level as a function of
-#' covariates with xgboost algorithm.
+#' covariates.
 #'
 #' @param cov_mt A covariate matrix containing all covariates. Each row is a
-#' sample and each column is a covariate.
+#' data sample and each column is a covariate.
 #' @param w_all A vector of observed exposure levels.
+#' @param sl_lib A vector of SuperLearner's package libraries.
 #' @param dnorm_log Logical, if TRUE, probabilities p are given as log(p).
 #'
 #' @return
@@ -19,40 +20,25 @@
 #' @export
 #'
 #' @examples
-#' mydata <- generate_synthetic_data()
-#' GPS_m <- train_GPS(mydata[,c("cf1", "cf2", "cf3", "cf4", "cf5", "cf6")],
-#'                    mydata$treat)
+#' data <- generate_synthetic_data()
+#' GPS_m <- train_gps(cov_mt = data[,-(1:2)],
+#'                   w_all = data$treat,
+#'                   sl_lib = c("SL.xgboost"),
+#'                   dnorm_log = FALSE)
 #'
 
-train_GPS <- function(cov_mt, w_all, dnorm_log = FALSE){
+train_gps <- function(cov_mt, w_all, sl_lib, dnorm_log){
 
   logger::log_info("Started estimating GPS values ... ")
   t_1 <- proc.time()
-  # GPS_fit <- CausalGPS::estimate_gps(NA, w_all,
-  #                                    cov_mt,
-  #                                    pred_model = "sl",
-  #                                    gps_model = "parametric",
-  #                                    internal_use = T,
-  #                                    params = list(xgb_max_depth = c(3,4,5),
-  #                                                  xgb_nrounds=c(50,60)),
-  #                                    nthread = 1,
-  #                                    sl_lib = c("m_xgboost", "m_ranger"))
-  #
-  # e_gps_pred <- GPS_fit$e_gps_pred
-  # e_gps_std <- GPS_fit$e_gps_std_pred
-  # GPS <- c(stats::dnorm(w_all, mean = e_gps_pred, sd = e_gps_std, log = dnorm_log))
-  # GPS_m <- data.frame(GPS = GPS,
-  #                                 e_gps_pred = e_gps_pred,
-  #                                 e_gps_std = e_gps_std)
-  GPS_SL <- SuperLearner::SuperLearner(Y = w_all, X = cov_mt,
-                                      SL.library = c("SL.earth",
-                                                     "SL.gam",
-                                                     "SL.glm",
-                                                     "SL.glm.interaction",
-                                                     "SL.mean",
-                                                     "SL.ranger"))
+  GPS_SL <- SuperLearner::SuperLearner(Y = w_all,
+                                       X = as.data.frame(cov_mt),
+                                       SL.library = sl_lib)
   GPS_SL_sd <- sd(w_all - GPS_SL$SL.predict)
-  GPS_m <- data.frame(GPS = dnorm(w_all, mean = GPS_SL$SL.predict, sd = GPS_SL_sd, log = dnorm_log),
+  GPS_m <- data.frame(GPS = dnorm(w_all,
+                                  mean = GPS_SL$SL.predict,
+                                  sd = GPS_SL_sd,
+                                  log = dnorm_log),
                       e_gps_pred = GPS_SL$SL.predict,
                       e_gps_std = GPS_SL_sd)