+++ tests/testthat/test-compute_deriv_nn.R
@@ -2,8 +2,10 @@ test_that("compute_deriv_nn works as expected!", {
 
   set.seed(765)
   data <- generate_synthetic_data(sample_size = 200)
-  GPS_m <- train_GPS(cov_mt = as.data.frame(data[,-(1:2)]),
-                     w_all = data$treat)
+  GPS_m <- train_gps(cov_mt = data[,-(1:2)],
+                     w_all = data$treat,
+                     sl_lib = c("SL.xgboost"),
+                     dnorm_log = FALSE)
 
   wi <- 4.8
 
@@ -17,7 +19,7 @@ test_that("compute_deriv_nn works as expected!", {
                                 block_size = 1000)
 
   expect_equal(length(deriv_val), 1L)
-  expect_equal(deriv_val[1,1], 0, tolerance = 0.00001)
+  expect_equal(deriv_val[1,1], -0.1793515, tolerance = 0.00001)
   expect_true(is.matrix(deriv_val))
 
 })