+++ tests/testthat/test-compute_deriv_weights_gp.R
@@ -2,8 +2,10 @@ test_that("compute_deriv_weights_gp works as expected!", {
 
   set.seed(9615)
   data <- generate_synthetic_data(sample_size = 200)
-  GPS_m <- train_GPS(cov_mt = data[,-(1:2)],
-                     w_all = data$treat)
+  GPS_m <- train_gps(cov_mt = data[,-(1:2)],
+                     w_all = data$treat,
+                     sl_lib = c("SL.xgboost"),
+                     dnorm_log = FALSE)
 
   wi <- 4.2
   weights <- compute_deriv_weights_gp(w = wi,
@@ -12,5 +14,5 @@ test_that("compute_deriv_weights_gp works as expected!", {
                                       hyperparam = c(1,1,2))
 
   expect_equal(length(weights), nrow(data))
-  # expect_equal(weights[37], -1.746696e-05, tolerance = 0.00001)
+  expect_equal(weights[37], 0.000708717, tolerance = 0.00001)
 })