+++ tests/testthat/test-compute_sd_gp.R
@@ -12,8 +12,10 @@ test_that("compute_sd_gp works as expected.", {
    kernel_fn <- function(x) exp(-x^2)
 
    # Estimate GPS function
-   GPS_m <- train_GPS(cov_mt = data[,-(1:2)],
-                      w_all = data$treat)
+   GPS_m <- train_gps(cov_mt = data[,-(1:2)],
+                      w_all = data$treat,
+                      sl_lib = c("SL.xgboost"),
+                      dnorm_log = FALSE)
 
    GPS <- GPS_m$GPS