+++ tests/testthat/test-estimate_cerf_gp.R
@@ -1,18 +1,19 @@
 test_that("estimate_cerf_gp works as expected!", {
 
   set.seed(129)
-  sim.data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)
+  data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)
 
   # Estimate GPS function
-  # In the future, CausalGPS gps estimation will be used.
-  GPS_m <- train_GPS(cov_mt = sim.data[,-(1:2)],
-                     w_all = sim.data$treat)
+  GPS_m <- train_gps(cov_mt = data[,-(1:2)],
+                     w_all = data$treat,
+                     sl_lib = c("SL.xgboost"),
+                     dnorm_log = FALSE)
 
   # exposure values
-  w.all <- seq(0,20,0.1)
+  w_all <- seq(0,20,0.1)
 
-  cerf_gp_obj <- estimate_cerf_gp(data = sim.data,
-                                  w = w.all,
+  cerf_gp_obj <- estimate_cerf_gp(data = data,
+                                  w = w_all,
                                   GPS_m = GPS_m,
                                   params = list(alpha = c(0.1,0.2,0.4),
                                                 beta=0.2,