+++ tests/testthat/test-estimate_mean_sd_nn.R
@@ -4,8 +4,10 @@ test_that("estimate_mean_sd_nn works as expected!", {
   data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)
 
   # Estimate GPS function
-  GPS_m <- train_GPS(cov_mt = data[,-(1:2)],
-                     w_all = data$treat)
+  GPS_m <- train_gps(cov_mt = data[,-(1:2)],
+                     w_all = data$treat,
+                     sl_lib = c("SL.xgboost"),
+                     dnorm_log = FALSE)
 
   # Hyperparameter
   hyperparam <- c(0.1, 0.2, 1)